feat(calendar): split calendar and event model
